tag: make tag.num_items unsigned
There's no point in declaring num_items as a uint8_t, it doesn't save
any space, due to padding.  This allows us to lift the articial "255
items" limitation.

+++ src/tag.h
@@ -58,7 +58,7 @@ struct tag_item {
 struct tag {
 	int time;
 	struct tag_item **items;
-	uint8_t num_items;
+	unsigned num_items;
 };
 
 struct tag *tag_ape_load(const char *file);
